+++ Source/DafnyLanguageServer.Test/Util/ClientBasedLanguageServerTest.cs
@@ -123,7 +123,7 @@ public class ClientBasedLanguageServerTest : DafnyLanguageServerTestBase {
   }
 
   public async Task AssertNoVerificationStatusIsComing(TextDocumentItem documentItem, CancellationToken cancellationToken) {
-    foreach (var entry in Documents.Documents.Values) {
+    foreach (var entry in Documents.Documents) {
       try {
         await entry.LastDocument;
       } catch (TaskCanceledException) {
@@ -140,7 +140,7 @@ public class ClientBasedLanguageServerTest : DafnyLanguageServerTestBase {
   }
 
   public async Task AssertNoGhostnessIsComing(CancellationToken cancellationToken) {
-    foreach (var entry in Documents.Documents.Values) {
+    foreach (var entry in Documents.Documents) {
       try {
         await entry.LastDocument;
       } catch (TaskCanceledException) {
@@ -162,7 +162,7 @@ public class ClientBasedLanguageServerTest : DafnyLanguageServerTestBase {
   }
 
   public async Task AssertNoDiagnosticsAreComing(CancellationToken cancellationToken) {
-    foreach (var entry in Documents.Documents.Values) {
+    foreach (var entry in Documents.Documents) {
       try {
         await entry.LastDocument;
       } catch (TaskCanceledException) {
