+++ Source/DafnyLanguageServer/Workspace/CompilationManager.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Data;
 using System.Linq;
 using System.Reactive.Concurrency;
 using System.Reactive.Linq;
@@ -10,7 +9,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Boogie;
 using Microsoft.Dafny.LanguageServer.Language;
-using Microsoft.Dafny.LanguageServer.Workspace.ChangeProcessors;
 using Microsoft.Dafny.LanguageServer.Workspace.Notifications;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -19,6 +17,10 @@ using VC;
 
 namespace Microsoft.Dafny.LanguageServer.Workspace;
 
+/// <summary>
+/// Manages the compilation of a single document version.
+/// Agnostic to document updates. Does not handle migration of old document state.
+/// </summary>
 public class CompilationManager {
 
   private readonly ILogger logger;
@@ -158,97 +160,98 @@ public class CompilationManager {
     NotifyStatus(translatedDocument.TextDocumentItem, lastDocument, cancellationSource.Token);
   }
 
-    public bool Verify(DafnyDocument dafnyDocument, IImplementationTask implementationTask) {
+  public bool Verify(DafnyDocument dafnyDocument, IImplementationTask implementationTask) {
 
-      var statusUpdates = implementationTask.TryRun();
-      if (statusUpdates == null) {
-        if (VerifierOptions.GutterStatus && implementationTask.CacheStatus is Completed completedCache) {
-          foreach (var result in completedCache.Result.VCResults) {
-            dafnyDocument.GutterProgressReporter!.ReportAssertionBatchResult(
-              new AssertionBatchResult(implementationTask.Implementation, result));
-          }
-
-          dafnyDocument.GutterProgressReporter!.ReportEndVerifyImplementation(implementationTask.Implementation,
-            completedCache.Result);
+    var statusUpdates = implementationTask.TryRun();
+    if (statusUpdates == null) {
+      if (VerifierOptions.GutterStatus && implementationTask.CacheStatus is Completed completedCache) {
+        foreach (var result in completedCache.Result.VCResults) {
+          dafnyDocument.GutterProgressReporter!.ReportAssertionBatchResult(
+            new AssertionBatchResult(implementationTask.Implementation, result));
         }
 
-        return false;
+        dafnyDocument.GutterProgressReporter!.ReportEndVerifyImplementation(implementationTask.Implementation,
+          completedCache.Result);
       }
 
-      MarkVerificationStarted();
-      statusUpdates.ObserveOn(verificationUpdateScheduler).Subscribe(
-        update => HandleStatusUpdate(dafnyDocument, implementationTask, update),
-        () => {
-          if (dafnyDocument.VerificationTasks!.All(t => t.IsIdle)) {
-            MarkVerificationFinished();
-          }
-        });
-
-      return true;
+      return false;
     }
 
-    private void HandleStatusUpdate(DafnyDocument document, IImplementationTask implementationTask, IVerificationStatus boogieStatus) {
-      var id = TextDocumentLoader.GetImplementationId(implementationTask.Implementation);
-      var status = StatusFromBoogieStatus(boogieStatus);
-      var implementationRange = implementationTask.Implementation.tok.GetLspRange();
-      if (boogieStatus is Running) {
-        if (VerifierOptions.GutterStatus) {
-          document.GutterProgressReporter!.ReportVerifyImplementationRunning(implementationTask.Implementation);
+    MarkVerificationStarted();
+    statusUpdates.ObserveOn(verificationUpdateScheduler).Subscribe(
+      update => HandleStatusUpdate(dafnyDocument, implementationTask, update),
+      () => {
+        if (dafnyDocument.VerificationTasks!.All(t => t.IsIdle)) {
+          MarkVerificationFinished();
         }
-      }
+      });
 
-      if (boogieStatus is Completed completed) {
-        var verificationResult = completed.Result;
-        foreach (var counterExample in verificationResult.Errors) {
-          document.Counterexamples!.Push(counterExample);
-        }
+    return true;
+  }
 
-        var diagnostics = GetDiagnosticsFromResult(document, verificationResult);
-        var view = new ImplementationView(implementationRange, status, diagnostics);
-        document.ImplementationIdToView!.AddOrUpdate(id, view, (_, _) => view);
-        if (VerifierOptions.GutterStatus) {
-          document.GutterProgressReporter!.ReportEndVerifyImplementation(implementationTask.Implementation, verificationResult);
-        }
-      } else {
-        document.ImplementationIdToView!.AddOrUpdate(id,
-          _ => new ImplementationView(implementationRange, status, Array.Empty<Diagnostic>()),
-          (_, previousView) => previousView with { Status = status });
-      }
-
-      documentUpdates.OnNext(document.Snapshot());
-    }
-
-    private List<Diagnostic> GetDiagnosticsFromResult(DafnyDocument document, VerificationResult result) {
-      var errorReporter = new DiagnosticErrorReporter(document.TextDocumentItem.Text, document.Uri);
-      foreach (var counterExample in result.Errors) {
-        errorReporter.ReportBoogieError(counterExample.CreateErrorInformation(result.Outcome, Options.ForceBplErrors));
-      }
-
-      var outcomeError = result.GetOutcomeError(Options);
-      if (outcomeError != null) {
-        errorReporter.ReportBoogieError(outcomeError);
-      }
-
-      return errorReporter.GetDiagnostics(document.Uri).OrderBy(d => d.Range.Start).ToList();
-    }
-
-    private PublishedVerificationStatus StatusFromBoogieStatus(IVerificationStatus verificationStatus) {
-      switch (verificationStatus) {
-        case Stale:
-          return PublishedVerificationStatus.Stale;
-        case Queued:
-          return PublishedVerificationStatus.Queued;
-        case Running:
-          return PublishedVerificationStatus.Running;
-        case Completed completed:
-          return completed.Result.Outcome == ConditionGeneration.Outcome.Correct
-            ? PublishedVerificationStatus.Correct
-            : PublishedVerificationStatus.Error;
-        default:
-          throw new ArgumentOutOfRangeException();
+  private void HandleStatusUpdate(DafnyDocument document, IImplementationTask implementationTask, IVerificationStatus boogieStatus) {
+    var id = TextDocumentLoader.GetImplementationId(implementationTask.Implementation);
+    var status = StatusFromBoogieStatus(boogieStatus);
+    var implementationRange = implementationTask.Implementation.tok.GetLspRange();
+    if (boogieStatus is Running) {
+      if (VerifierOptions.GutterStatus) {
+        document.GutterProgressReporter!.ReportVerifyImplementationRunning(implementationTask.Implementation);
       }
     }
 
+    if (boogieStatus is Completed completed) {
+      var verificationResult = completed.Result;
+      foreach (var counterExample in verificationResult.Errors) {
+        document.Counterexamples!.Push(counterExample);
+      }
+
+      var diagnostics = GetDiagnosticsFromResult(document, verificationResult);
+      var view = new ImplementationView(implementationRange, status, diagnostics);
+      document.ImplementationIdToView!.AddOrUpdate(id, view, (_, _) => view);
+      if (VerifierOptions.GutterStatus) {
+        document.GutterProgressReporter!.ReportEndVerifyImplementation(implementationTask.Implementation, verificationResult);
+      }
+    } else {
+      document.ImplementationIdToView!.AddOrUpdate(id,
+        _ => new ImplementationView(implementationRange, status, Array.Empty<Diagnostic>()),
+        (_, previousView) => previousView with { Status = status });
+    }
+
+    documentUpdates.OnNext(document.Snapshot());
+  }
+
+  private List<Diagnostic> GetDiagnosticsFromResult(DafnyDocument document, VerificationResult result) {
+    var errorReporter = new DiagnosticErrorReporter(document.TextDocumentItem.Text, document.Uri);
+    foreach (var counterExample in result.Errors) {
+      errorReporter.ReportBoogieError(counterExample.CreateErrorInformation(result.Outcome, Options.ForceBplErrors));
+    }
+
+    var outcomeError = result.GetOutcomeError(Options);
+    if (outcomeError != null) {
+      errorReporter.ReportBoogieError(outcomeError);
+    }
+
+    return errorReporter.GetDiagnostics(document.Uri).OrderBy(d => d.Range.Start).ToList();
+  }
+
+  private PublishedVerificationStatus StatusFromBoogieStatus(IVerificationStatus verificationStatus) {
+    switch (verificationStatus) {
+      case Stale:
+        return PublishedVerificationStatus.Stale;
+      case Queued:
+        return PublishedVerificationStatus.Queued;
+      case Running:
+        return PublishedVerificationStatus.Running;
+      case Completed completed:
+        return completed.Result.Outcome == ConditionGeneration.Outcome.Correct
+          ? PublishedVerificationStatus.Correct
+          : PublishedVerificationStatus.Error;
+      default:
+        throw new ArgumentOutOfRangeException();
+    }
+  }
+
+  // TODO migrate to higher level?
   private void NotifyStatus(TextDocumentItem item, DafnyDocument document, CancellationToken cancellationToken) {
     var errorCount = document.ImplementationIdToView!.Values.Sum(r => r.Diagnostics.Count(d => d.Severity == DiagnosticSeverity.Error));
     logger.LogDebug($"Finished verification with {errorCount} errors.");
@@ -278,6 +281,7 @@ public class CompilationManager {
     verificationCompleted.TrySetResult();
   }
 
+  // TODO migrate to higher level?
   public Task<DafnyDocument> LastDocument => TranslatedDocument.ContinueWith(t => {
     if (t.IsCompletedSuccessfully) {
 #pragma warning disable VSTHRD003
