+++ Source/DafnyLanguageServer/Workspace/LanguageServerExtensions.cs
@@ -27,7 +27,7 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
       return services
         .Configure<DocumentOptions>(configuration.GetSection(DocumentOptions.Section))
         .Configure<DafnyPluginsOptions>(configuration.GetSection(DafnyPluginsOptions.Section))
-        .AddSingleton<IDocumentDatabase>(serviceProvider => new DocumentDatabase(serviceProvider,
+        .AddSingleton<IDocumentDatabase>(serviceProvider => new DocumentManagerDatabase(serviceProvider,
           serviceProvider.GetRequiredService<IOptions<DocumentOptions>>().Value,
           serviceProvider.GetRequiredService<IOptions<VerifierOptions>>().Value))
         .AddSingleton<IDafnyParser>(serviceProvider => DafnyLangParser.Create(serviceProvider.GetRequiredService<ILogger<DafnyLangParser>>()))
